tools/requirements.txt
python-pptx>=0.6.21
Pillow>=9.0.0
pytesseract>=0.3.10
pdfminer.six>=20220524

# Note: For image extraction and OCR you may need poppler (pdftoppm) and Tesseract installed on the system.
